requirements.txt
discord.py
python-dotenv
psutil
gpiozero
